toy/requirements.txt
lightning
lightning[pytorch-extra]
tensorboard
imageio
imageio[ffmpeg]
lmdb
tqdm
scipy
seaborn
matplotlib
range-coder
pypng
torchac
pytorch-msssim
torchmetrics[image]
wandb
einops
opencv-python